Makefile.PL: Add INSTALLDIRS key-value pair
In the course of working on https://github.com/Perl/perl5/issues/12722,
it became evident that the CPAN version of Config-Perl-V's `Makefile.PL`
does not set an `INSTALLDIRS` key-value pair.  That raises the
possibility that if someone were to install the library against a
pre-5.12 perl, it would be installed in the wrong location.

This pull request adds an `INSTALLDIRS` key-value pair per the
recommendation in
https://github.com/Perl/perl5/issues/12722#issuecomment-544040160.
Please review.